+++ .env-example
@@ -1,5 +1,5 @@
 AUTH_SECRET=AUTH_SECRET
-NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY
+CLERK_PUBLISHABLE_KEY=CLERK_PUBLISHABLE_KEY
 CLERK_SECRET_KEY=CLERK_SECRET_KEY
 
 NEXT_PUBLIC_CLERK_SIGN_IN_URL=/sign-in
