+++ src/app/layout.tsx
@@ -10,9 +10,9 @@ import { SpeedInsights as VercelSpeedInsights } from "@vercel/speed-insights/nex
 import { auth } from "@clerk/nextjs/server";
 import { ClerkProvider } from "@clerk/nextjs";
 
-import { env } from "@/env.mjs";
+import { env } from "@/env.ts";
 import { siteConfig } from "~/config/site";
-import { ChildrenProps } from "~/src/types";
+import { ChildrenProps } from "~/types";
 
 import Providers from "~/components/providers";
 import Footer from "~/components/footer";
