+++ src/app/organizations/[org_slug]/tournaments/[tournament_id]/(tournament_page)/page.tsx
@@ -1,7 +1,7 @@
 import { getTournament, getTournaments } from "~/app/server-actions/tournaments/actions";
 import { Chip, Divider, Spacer } from "~/components/nextui/client-components";
 import OrganizationHeader from "~/components/organizations/org-header";
-import { OrganizationTournamentProps } from "~/src/types";
+import { OrganizationTournamentProps } from "~/types";
 
 import Link from "next/link";
