+++ src/components/miscellaneous/delete.tsx
@@ -1,4 +1,4 @@
-import type { IconSvgProps } from "~/src/types";
+import type { IconSvgProps } from "~/types";
 
 export const DeleteFilledIcon = (props: IconSvgProps) => (
   <svg
