+++ src/lib/api/client.ts
@@ -1,5 +1,5 @@
 import createFetchClient, { Middleware } from "openapi-fetch";
-import { env } from "@/env.mjs";
+import { env } from "@/env.ts";
 import { getVercelOidcToken } from "@vercel/functions/oidc";
 import { auth } from "@clerk/nextjs/server";
 import { paths } from "~/lib/api/openapi-v1";
