+++ src/lib/auth/cookies.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import * as cookie from "cookie";
-import { env } from "@/env.mjs";
+import { env } from "@/env.ts";
 import { headers } from "next/headers";
 
 function getCookieDomain() {
