+++ src/middleware.ts
@@ -1,5 +1,5 @@
 import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
-import { env } from "@/env.mjs";
+import { env } from "@/env.ts";
 
 export const config = {
   matcher: [
